requirements.txt
flask==2.3.3
flask-cors==4.0.0
paho-mqtt==1.6.1
